refactor : 코딩 컨벤션 준수를 위한 코드 정리
