+++ src/main/java/gift/controller/AdminController.java
@@ -1,7 +1,7 @@
-package gift.Controller;
+package gift.controller;
 
-import gift.Model.Product;
-import gift.Service.ProductService;
+import gift.model.Product;
+import gift.service.ProductService;
 import jakarta.validation.Valid;
 import java.util.List;
 import org.springframework.stereotype.Controller;
