+++ src/main/java/gift/controller/LoginDTO.java
@@ -1,4 +1,4 @@
-package gift.Controller;
+package gift.controller;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -10,4 +10,6 @@ public record LoginDTO(
 
     @NotBlank(message = "비밀번호를 입력하세요.")
     String password
-) {}
+) {
+
+}
