+++ src/main/java/gift/controller/UserDTO.java
@@ -1,4 +1,4 @@
-package gift.Controller;
+package gift.controller;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -13,4 +13,6 @@ public record UserDTO(
 
     @NotBlank(message = "비밀번호를 입력하세요.")
     String password
-) {}
+) {
+
+}
