+++ src/main/java/gift/model/Product.java
@@ -1,4 +1,4 @@
-package gift.Model;
+package gift.model;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
@@ -16,4 +16,5 @@ public record Product(Long id,
                       @NotBlank(message = "이미지 URL을 입력해야 합니다.")
                       @Pattern(regexp = "^(http|https)://.*$", message = "유효한 이미지 URL을 입력해야 합니다.")
                       String imageUrl) {
+
 }
