+++ src/main/java/gift/model/User.java
@@ -1,4 +1,4 @@
-package gift.Model;
+package gift.model;
 
 public record User(Long id, String name, String email, String password, String role) {
