+++ src/main/resources/db/data.sql
@@ -2,6 +2,9 @@ INSERT INTO products (name, price, imageUrl)
 VALUES ('예시1', 100, '예시1 Image Url'),
        ('예시2', 200, '예시2 Image Url');
 
-INSERT INTO users (name, email, password, role) VALUES ('관리자','admin@email.com', 'password', 'admin');
-INSERT INTO users (name, email, password, role) VALUES ('사용자','user@email.com', 'password', 'user');
-INSERT INTO users (name, email, password, role) VALUES ('카카오','kakaouser@email.com', 'password', 'kakaouser');
+INSERT INTO users (name, email, password, role)
+VALUES ('관리자', 'admin@email.com', 'password', 'admin');
+INSERT INTO users (name, email, password, role)
+VALUES ('사용자', 'user@email.com', 'password', 'user');
+INSERT INTO users (name, email, password, role)
+VALUES ('카카오', 'kakaouser@email.com', 'password', 'kakaouser');
