+++ src/main/resources/db/schema.sql
@@ -8,9 +8,9 @@ CREATE TABLE products
 
 CREATE TABLE users
 (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    name VARCHAR(255) NOT NULL,
-    email VARCHAR(255) NOT NULL UNIQUE,
+    id       BIGINT AUTO_INCREMENT PRIMARY KEY,
+    name     VARCHAR(255) NOT NULL,
+    email    VARCHAR(255) NOT NULL UNIQUE,
     password VARCHAR(255) NOT NULL,
-    role VARCHAR(50) NOT NULL
+    role     VARCHAR(50)  NOT NULL
 );
